+++ src/entities/Entity.cpp
@@ -1456,7 +1456,7 @@ SpritePtr Entity::create_sprite(
     int order
 ) {
   if (order == -1) {
-    order = sprites.size();
+    order = static_cast<int>(sprites.size());
   }
   SpritePtr sprite = Sprite::create(animation_set_id);
   SOLARUS_REQUIRE(sprite, "Entity::create_sprite failed.");
